Refactor/#78: ReactionType enum 클래스 위치 이동
