--- cluvr-api/src/main/java/com/example/cluvrapi/domain/board/enums/ReactionType.java
@@ -1,5 +1,0 @@
-package com.example.cluvrapi.domain.board.enums;
-
-public enum ReactionType {
-	LIKE, DISLIKE
-}
