+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/reaction/enums/ReactionType.java
@@ -1,0 +1,5 @@
+package com.example.cluvrapi.domain.reaction.enums;
+
+public enum ReactionType {
+	LIKE, DISLIKE
+}
